requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
numpy<2.0.0
torch==2.2.2+cpu
torchvision==0.17.2+cpu
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-groq==0.2.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0
docling==2.15.1
gradio==4.44.0
python-dotenv==1.0.1
rank-bm25==0.2.2